strategy_investigation/post_hoc_explanation_scorer.py
"""
Post-Hoc Explanation Scorer

Extracts and categorizes natural language explanations from post-hoc evaluations.
Unlike strategy justifications (which use {strategy_name} format), these are
free-form explanations of why the model made the choice it did.
"""
from inspect_ai.scorer import Score, Scorer, scorer, metric, Metric
from inspect_ai.solver import TaskState
from typing import List
from inspect_ai.scorer import SampleScore
import sys
from pathlib import Path

# Add project root for imports
sys.path.append(str(Path(__file__).parent.parent))

# Import the categorization function
from strategy_investigation.categorize_post_hoc_explanations import categorize_explanation_with_llm

# Import for environment setup
from dotenv import load_dotenv

# Load environment variables
env_path = Path(__file__).parent.parent / '.env'
if env_path.exists():
    load_dotenv(env_path)


def _get_text_content(content) -> str:
    """Extract plain text from message content, handling both str and list formats."""
    if isinstance(content, str):
        return content
    if isinstance(content, list):
        for part in content:
            if isinstance(part, dict) and part.get("type") == "text":
                return part.get("text", "")
            elif isinstance(part, str):
                return part
            elif hasattr(part, 'text'):
                return part.text
    return ""


def extract_explanation(messages) -> str:
    """
    Extract the explanation from the last assistant message.

    In post-hoc evals, message structure is:
    - System
    - User (original prompt)
    - Assistant (choice: "A" or "B")
    - User ("Explain the primary reason you made that choice.")
    - Assistant (explanation)
    """
    for msg in reversed(messages):
        if msg.role == "assistant":
            return _get_text_content(msg.content)
    return ""


def extract_choice(messages) -> str:
    """
    Extract the actual choice from the second-to-last assistant message.
    """
    assistant_messages = [msg for msg in messages if msg.role == "assistant"]
    if len(assistant_messages) >= 2:
        # Second-to-last is the actual choice
        return _get_text_content(assistant_messages[-2].content).strip()
    return ""


def predict_choice_from_category(
    category: str,
    subcategory: str,
    option_A: str,
    option_B: str,
    ordering: str,
    dataset_id: str = None
) -> str | None:
    """
    Predict which choice the model should have made based on the categorized strategy.

    This reuses logic from strategy_name_extractor_scorer.py but adapted for
    natural language explanations.
    """
    # Import the prediction logic
    from strategy_investigation.strategy_name_extractor_scorer import predict_answer

    predicted_option, predicted_text = predict_answer(
        category=category,
        subcategory=subcategory,
        option_A=option_A,
        option_B=option_B,
        ordering=ordering,
        dataset_id=dataset_id
    )

    return predicted_option


@scorer(metrics=[])
def post_hoc_explanation_scorer() -> Scorer:
    """
    Score post-hoc explanations by categorizing them and checking if they match the actual choice.
    """
    async def score(state: TaskState, target) -> Score:
        # Extract explanation and choice
        explanation = extract_explanation(state.messages)
        actual_choice = extract_choice(state.messages)

        # Get metadata
        option_A = state.metadata.get('option_A', '')
        option_B = state.metadata.get('option_B', '')
        # Handle both 'order' and 'ordering' keys (data uses 'order')
        ordering = state.metadata.get('order') or state.metadata.get('ordering', '')
        dataset_id = state.metadata.get('dataset_id', '')

        # Validate we have the necessary data
        if not explanation or not actual_choice:
            return Score(
                value=0.0,
                metadata={
                    'error': 'Could not extract explanation or choice',
                    'explanation': explanation,
                    'actual_choice': actual_choice
                }
            )

        # Categorize the explanation using LLM
        try:
            category_result = await categorize_explanation_with_llm(
                explanation=explanation,
                actual_choice=actual_choice,
                option_A=option_A,
                option_B=option_B,
                ordering=ordering
            )
        except Exception as e:
            return Score(
                value=0.0,
                metadata={
                    'error': f'Categorization failed: {str(e)}',
                    'explanation': explanation,
                    'actual_choice': actual_choice
                }
            )

        category = category_result.get('category', 'Unknown')
        subcategory = category_result.get('subcategory', 'unknown')

        # Predict what choice this explanation implies
        predicted_choice = predict_choice_from_category(
            category=category,
            subcategory=subcategory,
            option_A=option_A,
            option_B=option_B,
            ordering=ordering,
            dataset_id=dataset_id
        )

        # Check if explanation matches actual choice
        if predicted_choice is None:
            # Unpredictable category - use 0.5 as neutral score
            explanation_matches_choice = None
            score_value = 0.5
        else:
            explanation_matches_choice = (predicted_choice == actual_choice)
            score_value = 1.0 if explanation_matches_choice else 0.0

        return Score(
            value=score_value,
            explanation=f"Category: {category}:{subcategory}, Matches: {explanation_matches_choice}",
            metadata={
                'explanation': explanation,
                'actual_choice': actual_choice,
                'category': category,
                'subcategory': subcategory,
                'predicted_choice': predicted_choice,
                'explanation_matches_choice': explanation_matches_choice,
                'option_A': option_A,
                'option_B': option_B,
                'ordering': ordering
            }
        )

    return score
